+++ ui.R
@@ -24,7 +24,8 @@ ui <- shiny::tagList(
         tags$script(src = "https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.13/js/select2.min.js")
       ),
       loadModuleUIs(appConfig),
-      shinyjs::useShinyjs()
+      shinyjs::useShinyjs(),
+      waiter::useWaiter()
     )
   ),
   tags$footer(footer(getAppId(), getAppVersion(), getAppDate()), align = "center")